requirements.txt
Django==3.0.2
json-rpc==1.10.8
requests==2.18.4
Werkzeug==0.15.2
PyYAML==5.3
beautifulsoup4==4.8.2
coverage==5.0.3